Решена проблема с сильными ссылками, создана фабрика вопросов, соответствующий протокол
